Create an `ItemDecorations` file for all `RecyclerView` item decorations
All item decorations are now placed in one file so that any `RecyclerView` in the app can use the same ones.
